feat(severity-indicator): grouped timeline logic into single logic.ts file as all individual utilities are very small yet depent on each other; cleaned up JSX a bit;
