--- packages/app/src/components/severity-indicator-tile/components/timeline/logic/get-timeline-range-dates.ts
@@ -1,15 +1,0 @@
-import { createDateFromUnixTimestamp } from '~/utils/create-date-from-unix-timestamp';
-import { SeverityIndicatorTimelineEventConfig } from '../timeline';
-
-export const getTimelineRangeDates = (timelineEvents: SeverityIndicatorTimelineEventConfig[]) => {
-  const timelineEventDates = timelineEvents
-    .flatMap((timelineEvent: SeverityIndicatorTimelineEventConfig) => [timelineEvent.start, timelineEvent.end])
-    .map((timelineEventDate: number) => {
-      return createDateFromUnixTimestamp(timelineEventDate).getTime();
-    });
-
-  return {
-    startDate: Math.min(...timelineEventDates),
-    endDate: Math.max(...timelineEventDates),
-  };
-};
